feat(daemon): cross-platform support for Windows
Implements Windows equivalents for the POSIX-only daemon primitives so
aileron and aileron-server compile and run on Windows. Closes #574.

discovery package:
- Lock: flock(2) on POSIX → LockFileEx (LOCKFILE_EXCLUSIVE_LOCK |
  LOCKFILE_FAIL_IMMEDIATELY) on Windows. Same non-blocking semantic.
- LockFileInode: syscall.Stat_t.Ino on POSIX →
  GetFileInformationByHandle FileIndex (high<<32|low) on Windows.
  Stable across calls; returns wrapped os.ErrNotExist when file is
  absent, matching POSIX contract.

spawn package:
- detachProcAttrs(): SysProcAttr{Setsid:true} on POSIX →
  SysProcAttr{CreationFlags:DETACHED_PROCESS|CREATE_NEW_PROCESS_GROUP}
  on Windows. Disconnects daemon from parent's console / signal group
  so it survives the parent's exit.

cmd/aileron:
- signalDaemonStop(pid): syscall.Kill SIGTERM on POSIX →
  os.Process.Kill (TerminateProcess) on Windows. Windows kill is
  abrupt — daemon's shutdown defer doesn't run, so daemon.json may be
  left stale; the existing PID-not-alive cleanup path handles that on
  next stop.

Tests:
- POSIX-specific tests (chmod-based, unlink-while-locked) moved to
  *_unix_test.go with //go:build unix. Behavior pinned by these tests
  doesn't apply on Windows: chmod 0500 is no-op on dirs, and
  LockFileEx prevents unlinking a held file.
- New *_windows_test.go covers the Windows code paths: file-ID
  stability, missing-file os.ErrNotExist contract, and the
  remove-while-held block (the Windows-specific behavior that
  obsoletes the POSIX orphaned-inode failure mode from #528 3b).
- TestTailSessionLog_OpenFailureSurfaced gains a runtime Windows
  skip — 0o000 dir mode doesn't restrict open() on Windows so the
  error path under test is unreachable there.

CI:
- New test-windows job runs the platform-specific test packages on
  windows-latest. The full suite still runs on ubuntu-latest. Targeted
  scope keeps Windows runner cost down while validating the new code.

Validated locally:
- darwin/linux/windows builds clean for internal/, cmd/aileron, and
  cmd/aileron-mcp (cross-compile).
- goreleaser snapshot run produces the full matrix in 11s; previously
  failed on windows_amd64 for aileron-server and aileron.
- Existing test suites pass on darwin.

Refs #572 — unblocks cutting v0.0.1 with Windows binaries in the
release matrix.
