+++ .github/workflows/ci.yml
@@ -105,6 +105,33 @@ jobs:
         with:
           token: ${{ secrets.CODECOV_TOKEN }}
 
+  test-windows:
+    # Targeted Windows job: exercises the platform-specific code paths
+    # added for Windows support (#574) — discovery (LockFileEx,
+    # GetFileInformationByHandle), spawn (DETACHED_PROCESS), and the
+    # aileron CLI's stop helper (TerminateProcess). The full test suite
+    # still runs on ubuntu-latest above.
+    name: Unit Tests (Windows)
+    runs-on: windows-latest
+    steps:
+      - uses: actions/checkout@v5
+
+      - uses: actions/setup-go@v6
+        with:
+          go-version: "1.25"
+          cache-dependency-path: |
+            internal/go.mod
+            cmd/aileron/go.mod
+            cmd/aileron-mcp/go.mod
+            sdk/go/go.mod
+
+      - name: Run platform-specific tests
+        shell: bash
+        run: |
+          set -e
+          cd internal && go test -race ./daemon/discovery/... ./daemon/spawn/...
+          cd ../cmd/aileron && go test -race ./...
+
   test-ui:
     name: UI Tests
     runs-on: ubuntu-latest
